Replace Podman with Docker for container isolation

+++ getamp.sh
@@ -118,6 +118,7 @@ TPUT_IS_PRESENT="$(isPresent tput)"
 SELINUX_IS_INSTALLED="$(isPresent setsebool)"
 PODMAN_IS_INSTALLED="$(isPresent podman)"
 UIDMAP_IS_INSTALLED="$(isPresent newuidmap)"
+DOCKER_IS_INSTALLED="$(isPresent docker)"
 APT_IS_PRESENT="$(isPresent apt-get)"
 YUM_IS_PRESENT="$(isPresent yum)"
 PACMAN_IS_PRESENT="$(isPresent pacman)"
@@ -266,7 +267,7 @@ fi
 
 if [ "$ID" == "photon" ]; then
 	PREREQ_PACKAGES="wget tmux socat unzip git bindutils tar jq sqlite-devel"
-	FORCE_PODMAN=1
+	FORCE_CONTAINERS=1
 fi
 
 if [ "$INSTALL_IN_PROGRESS" ]; then
@@ -538,8 +539,12 @@ function promptForAMPUser {
 }
 
 function promptForDeps {
-	if [ -n "$FORCE_PODMAN" ]; then
-		installPodman=y
+	if [ -n "$FORCE_CONTAINERS" ]; then
+		if awk '$1=="0" && $2!="0"' /proc/self/uid_map | grep -q .; then
+			installDocker=y
+		else
+			installPodman=y
+		fi
 		return; 
 	fi
 
@@ -547,18 +552,34 @@ function promptForDeps {
 		installJava=$ANSWER_INSTALLJAVA
 		install32BitLibs=${ANSWER_INSTALL32BITLIBS:-${ANSWER_INSTALLSRCDSLIBS:-}}
 		installPodman=$ANSWER_INSTALLPODMAN
+		installDocker=$ANSWER_INSTALLDOCKER
 		return
 	fi
 
-	echo "Would you like to isolate your AMP instances by running them inside Podman containers?"
-	prnt "This provides an additional layer of protection at the expense of a minor performance impact. It is strongly recommended if you are going to allow untrusted users access to AMP."
-	echo
-	prnt "Using Podman is also strongly recommended for running some applications, as it removes the requirement to install additional dependencies on the host."
-	prnt "If you are using a Desktop environment / GUI on this system, you should use this option to avoid package conflicts."
-	read -rp "[y/N] " installPodman
-	installPodman=${installPodman:-n}
-	echo
-	echo
+	if awk '$1=="0" && $2!="0"' /proc/self/uid_map | grep -q .; then
+		echo "Would you like to isolate your AMP instances by running them inside Docker containers?"
+		prnt "This provides an additional layer of protection. It is strongly recommended if you are going to allow untrusted users access to AMP."
+		prnt "Using Docker is also strongly recommended for running some applications, as it removes the requirement to install additional dependencies on the host."
+		echo
+		prnt "You are attempting to install AMP within an unprivileged container. It is strongly recommended that you run AMP within a proper VM when able."
+		prnt "AMP is unable to install rootless Podman in this environment due to security restraints in the OS. Docker will be installed in place of Podman."
+		prnt "While running Docker does provide additional security versus native, running Docker as root still poses security risks."
+		read -rp "[y/N] " installDocker
+		installDocker=${installDocker:-n}
+		echo
+		echo
+	else
+		echo "Would you like to isolate your AMP instances by running them inside Podman containers?"
+		prnt "This provides an additional layer of protection at the expense of a minor performance impact. It is strongly recommended if you are going to allow untrusted users access to AMP."
+		echo
+		prnt "Using Podman is also strongly recommended for running some applications, as it removes the requirement to install additional dependencies on the host."
+		prnt "If you are using a Desktop environment / GUI on this system, you should use this option to avoid package conflicts."
+		read -rp "[y/N] " installPodman
+		installPodman=${installPodman:-n}
+		echo
+		echo
+	fi
+
 
 	echo "Will you be running Minecraft servers on this installation?"
 	echo "If selected, this installs the required versions of Java."
@@ -803,6 +824,15 @@ function installPodman {
 		} &>> "$LOG_FILE"
 		return
 	fi
+	
+	if awk '$1=="0" && $2!="0"' /proc/self/uid_map | grep -q .; then
+		prnt "You are attempting to install AMP within an unprivileged container. It is strongly recommended that you run AMP within a proper VM when able."
+		prnt "AMP is unable to install rootless Podman in this environment due to security restraints in the OS. You can install Docker using the \"installDocker\" flag."
+		prnt "While running Docker does provide additional security versus native, running Docker as root still poses security risks."
+		echo
+		echo
+		return
+	fi
 
 	echo "Installing Podman..."
 	{
@@ -817,6 +847,112 @@ unqualified-search-registries = ["docker.io"]
 short-name-mode = "permissive"
 EOF
 	} &>> "$LOG_FILE"
+}
+
+function installDocker {
+	echo ""
+	
+	if [[ "$DOCKER_IS_INSTALLED" ]]; then
+		echo "Docker is already installed."
+		echo "If you didn't install Docker from the official Docker repositories, then it may not operate correctly with AMP."
+		echo "Do you want to remove the existing Docker installation and install Docker from the official Docker repositories, if available for your system?"
+		echo "This will also stop any existing running Docker containers."
+		read -rp "[y/N] " reInstallDocker
+		reInstallDocker=${reInstallDocker:-n}
+		if [[ ! "$reInstallDocker" =~ ^[Yy]$ ]]; then
+			echo "Skipping Docker re-installation..."
+			{
+				usermod -a -G docker $AMP_SYS_USER
+				systemctl enable docker
+				systemctl start docker
+			} &>> "$LOG_FILE"
+			return
+		fi
+	fi
+
+	echo "Installing Docker..."
+	IFS='|' read -r BASE_ID BASE_SUITE BASE_VERSION_ID < <(mapUpstream)
+	DOCKER_REPO_AVAILABLE=false
+
+    case "$BASE_ID" in
+        ubuntu|debian)
+			if wget -q --spider https://download.docker.com/linux/$BASE_ID/dists/$BASE_SUITE/ >/dev/null 2>&1; then
+				DOCKER_REPO_AVAILABLE=true
+				if [[ "$reInstallDocker" =~ ^[Yy]$ ]]; then
+					REMOVE_DOCKER_PACKAGES="docker.io docker-doc docker-compose docker-compose-v2 podman-docker containerd runc"
+				fi
+				[[ -f /usr/share/keyrings/download.docker.com.gpg ]] && rm -f /usr/share/keyrings/download.docker.com.gpg >/dev/null 2>&1
+				[[ -f /etc/apt/sources.list.d/download.docker.com.list ]] && rm -f /etc/apt/sources.list.d/download.docker.com.list >/dev/null 2>&1
+				[[ -f /etc/apt/sources.list.d/docker.list ]] && rm -f /etc/apt/sources.list.d/docker.list >/dev/null 2>&1
+				if { [[ "$BASE_ID" == "ubuntu" ]] && version_ge "$BASE_VERSION_ID" "22.04"; } || { [[ "$BASE_ID" == "debian" ]] && version_ge "$BASE_VERSION_ID" "12"; }; then
+					printf "Types: deb\nURIs: https://download.docker.com/linux/%s\nSuites: %s\nComponents: stable\nArchitectures: %s\nSigned-By: /usr/share/keyrings/docker.asc\n" "$BASE_ID" "$BASE_SUITE" "$(dpkg --print-architecture)" \
+					| tee /etc/apt/sources.list.d/docker.sources > /dev/null
+				else
+					echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker.asc] https://download.docker.com/linux/$BASE_ID $BASE_SUITE stable" \
+					| tee /etc/apt/sources.list.d/docker.list > /dev/null
+				fi
+				{
+					install -d -m 0755 /usr/share/keyrings
+					wget -qO /usr/share/keyrings/docker.asc https://download.docker.com/linux/$BASE_ID/gpg
+					chmod a+r /usr/share/keyrings/docker.asc
+					$PM_COMMAND update
+					DOCKER_PACKAGES="docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin"
+				} &>> "$LOG_FILE"
+			fi
+            ;;
+        rhel|fedora|centos)
+			if wget -q --spider https://download.docker.com/linux/$BASE_ID/ >/dev/null 2>&1; then
+				DOCKER_REPO_AVAILABLE=true
+				[[ -f /etc/yum.repos.d/docker-ce.repo ]] && rm -f /etc/yum.repos.d/docker-ce.repo >/dev/null 2>&1
+				if [[ "$reInstallDocker" =~ ^[Yy]$ ]]; then
+					REMOVE_DOCKER_PACKAGES="docker docker-client docker-client-latest docker-common docker-latest docker-latest-logrotate docker-logrotate docker-selinux docker-engine-selinux docker-engine podman runc"
+				fi
+				if [[ "$PM_COMMAND" == "yum" ]]; then
+					{
+						$PM_COMMAND "${PM_INSTALL[@]}" yum-utils
+						yum-config-manager --add-repo https://download.docker.com/linux/$BASE_ID/docker-ce.repo
+					} &>> "$LOG_FILE"
+				elif [[ "$PM_COMMAND" == "dnf" ]]; then
+					{
+						$PM_COMMAND "${PM_INSTALL[@]}" dnf-plugins-core
+						$PM_COMMAND config-manager --add-repo https://download.docker.com/linux/$BASE_ID/docker-ce.repo
+					} &>> "$LOG_FILE"
+				fi
+				DOCKER_PACKAGES="docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin"
+			fi
+            ;;
+        *) ;;
+    esac
+
+	if [[ ! $DOCKER_REPO_AVAILABLE ]]; then
+		if [[ "$DOCKER_IS_INSTALLED" ]]; then
+			echo "Automatic Docker re-installation is not supported on your system at this time. After setup completes, please investigate using your current Docker installation with AMP, and if necessary re-installing Docker manually from the official Docker sources. See https://docs.docker.com/engine/install/ for more information. If you re-install Docker manually, also ensure that the '$AMP_SYS_USER' user is in the 'docker' group."
+			echo "Continuing without re-installing Docker..."
+			{
+				usermod -a -G docker $AMP_SYS_USER
+				systemctl enable docker
+				systemctl start docker
+			} &>> "$LOG_FILE"
+			return
+		else
+			echo "Automatic Docker installation is not supported on your system at this time. Please investigate installing it manually after setup completes. See https://docs.docker.com/engine/install/ for more information. If you install Docker manually, also ensure that the '$AMP_SYS_USER' user is added to the 'docker' group."
+			echo "Continuing without installing Docker..."
+			PROVISIONFLAGS="${PROVISIONFLAGS/ +ADSModule.Defaults.UseDocker True/}"
+			return
+		fi
+	else
+		{
+			if [[ "$reInstallDocker" =~ ^[Yy]$ ]] && [[ -n "$REMOVE_DOCKER_PACKAGES" ]]; then
+				docker ps -q | xargs -r docker stop
+				systemctl stop docker
+				for pkg in $REMOVE_DOCKER_PACKAGES; do $PM_COMMAND "${PM_UNINSTALL[@]}" $pkg; done
+			fi
+			$PM_COMMAND "${PM_INSTALL[@]}" $DOCKER_PACKAGES
+			systemctl enable docker
+			systemctl start docker
+			usermod -a -G docker $AMP_SYS_USER
+		} &>> "$LOG_FILE"
+	fi
 }
 
 function install32BitDeps {
@@ -876,6 +1012,11 @@ function installDependencies {
 
 	if [[ "$installPodman" =~ ^[Yy]$ ]]; then
 		installPodman
+		PROVISIONFLAGS="$PROVISIONFLAGS +ADSModule.Defaults.UseDocker True"
+	fi
+
+	if [[ "$installDocker" =~ ^[Yy]$ ]]; then
+		installDocker
 		PROVISIONFLAGS="$PROVISIONFLAGS +ADSModule.Defaults.UseDocker True"
 	fi
 
@@ -1305,9 +1446,10 @@ echo -en "Instance Manager:\t\t"| tee -a $INSTALL_SUMMARY
 if [ "$AMPINSTMGR_IS_INSTALLED" ]; then echo "Already installed"; else echo "To be installed"; fi| tee -a $INSTALL_SUMMARY
 echo -en "HTTPS setup:\t\t\t"| tee -a $INSTALL_SUMMARY
 if [[ "$setupnginx" =~ ^[Yy]$ ]]; then echo "Yes, via nginx with domain $nginxdomain"; else echo "No"; fi| tee -a $INSTALL_SUMMARY
-noReason=$( [[ "$installPodman" =~ ^[Yy]$ ]] && echo "Not required (replying on Podman)" || echo "No" )
-echo -en "Install Podman:\t\t\t" | tee -a $INSTALL_SUMMARY
-if [[ "$installPodman" =~ ^[Yy]$ ]]; then echo "Yes"; else echo "No"; fi | tee -a $INSTALL_SUMMARY
+noReason=$(([[ "$installPodman" =~ ^[Yy]$ ]] || [[ "$installDocker" =~ ^[Yy]$ ]]) && echo "Not required (running in containers)" || echo "No" )
+if [[ "$installPodman" =~ ^[Yy]$ ]]; then echo -e "Install Podman:\t\t\tYes";
+elif [[ "$installDocker" =~ ^[Yy]$ ]]; then echo -e "Install Docker:\t\t\tYes";
+else echo -e "Install Podman/Docker:\t\tNo"; fi | tee -a $INSTALL_SUMMARY
 if [ "$ARCH" == "x86_64" ]; then
 	echo -en "Install 32-bit libraries:\t" | tee -a $INSTALL_SUMMARY
 	if [[ "$install32BitLibs" =~ ^[Yy]$ ]]; then echo "Yes"; else echo "$noReason"; fi | tee -a $INSTALL_SUMMARY
